refactor #118: move components to atoms or molecules folders
