+++ app/leaderboard/page.tsx
@@ -1,8 +1,8 @@
-import Error from '@/components/error'
+import Error from '@/components/molecules/error'
 import Table from '@/components/table'
 import { BestTime, PlayersStats } from '@/interfaces/player'
 import { ThemeProvider } from '@/hooks/useTheme'
-import Navbar from '@/components/navbar'
+import Navbar from '@/components/molecules/navbar'
 import Footer from '@/components/molecules/footer'
 import PageHeader from '@/components/pageHeader'
 import { statsColumns } from '@/constants'
