+++ app/stats/[slug]/page.tsx
@@ -1,13 +1,13 @@
 import Footer from '@/components/molecules/footer'
-import Navbar from '@/components/navbar'
+import Navbar from '@/components/molecules/navbar'
 import PageHeader from '@/components/pageHeader'
 import { statsColumns, timeAllDiffColumns } from '@/constants'
 import { ThemeProvider } from '@/hooks/useTheme'
 import { PlayersStats } from '@/interfaces/player'
 import { headers } from 'next/headers'
-import Error from '@/components/error'
+import Error from '@/components/molecules/error'
 import TableWithControls from '@/components/tableWithControls'
-import ScoreboardSelector from '@/components/scoreboardSelector'
+import ScoreboardSelector from '@/components/molecules/scoreboardSelector'
 import Info from '@/components/atoms/info'
 import Link from '@/components/atoms/link'
