+++ components/atoms/input/index.tsx
@@ -2,7 +2,7 @@
 
 import React, { ReactNode } from 'react'
 import { ClearIconContainer, StyledInput, Wrapper } from './styled'
-import { Cross } from '../icons/cross'
+import { Cross } from '../../icons/cross'
 import { useTheme } from '@/hooks/useTheme'
 
 interface Props {
