+++ components/molecules/error/index.tsx
@@ -1,12 +1,15 @@
 'use client'
 
-import { Button, Container } from './styled'
+import Link from '../../atoms/link'
+import { Container } from './styled'
 
 export default function Error() {
   return (
     <Container>
       <p>There was an error handling the request</p>
-      <Button href="/">Return to home</Button>
+      <Link href="/" withButtonStyle>
+        Return to home
+      </Link>
     </Container>
   )
 }
