+++ components/molecules/error/styled.ts
@@ -1,15 +1,4 @@
 import styled from 'styled-components'
-import Link from 'next/link'
-
-export const Button = styled(Link)`
-  background-color: ${({ theme }) => theme.color.primary};
-  color: ${({ theme }) => theme.text.white};
-  border: none;
-  border-radius: 3px;
-  cursor: pointer;
-  padding: 12px 18px;
-  text-decoration: none;
-`
 
 export const Container = styled.section`
   align-items: center;
