+++ components/playerFinder/index.tsx
@@ -3,7 +3,7 @@
 import React, { useState, useEffect, useCallback } from 'react'
 import { Options, Option, Wrapper } from './styled'
 import { PlayersStats, PlayerStats } from '@/interfaces/player'
-import Input from '../input'
+import Input from '../atoms/input'
 import { Search } from '../icons/search'
 import { useTheme } from '@/hooks/useTheme'
 import Image from 'next/image'
